+++ noodles-bgzf/src/writer.rs
@@ -89,19 +89,6 @@ impl<W> Writer<W>
 where
     W: Write,
 {
-    /// Creates a BGZF writer builder.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use noodles_bgzf as bgzf;
-    /// let builder = bgzf::Writer::builder(Vec::new());
-    /// let writer = builder.build();
-    /// ```
-    pub fn builder(inner: W) -> Builder<W> {
-        Builder::new(inner)
-    }
-
     /// Creates a writer with a default compression level.
     ///
     /// # Examples
@@ -111,7 +98,7 @@ where
     /// let writer = bgzf::Writer::new(Vec::new());
     /// ```
     pub fn new(inner: W) -> Self {
-        Self::builder(inner).build()
+        Builder::default().build_with_writer(inner)
     }
 
     /// Returns a reference to the underlying writer.